requirements.txt
anyio==3.6.2
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
fastapi-jwt-auth==0.5.0
flake8==5.0.4
greenlet==2.0.2
h11==0.14.0
idna==3.4
mccabe==0.7.0
passlib==1.7.4
psycopg2-binary==2.9.5
phonenumbers==8.13.11
pycodestyle==2.9.1
pydantic==1.10.7
pyflakes==2.5.0
PyJWT==1.7.1
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.13
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0
